Completed TODO functions in Model and added a bunch in DAOs and Data Access

+++ back_end/src/main/java/back/data/jdbc/RoomRowMapper.java
@@ -13,7 +13,7 @@ public class RoomRowMapper implements RowMapper<Room> {
 
     @Override
     public Room mapRow(ResultSet rs, int rowNum) throws SQLException{
-        long id = rs.getLong("id");
+        int id = rs.getInt("id");
         long provider_id = rs.getLong("provider_id");
         double price = rs.getDouble("price");
         int capacity = rs.getInt("capacity");
